file-xfer: use better names for duplicate files
When transferring the same file multiple times, the vdagent tries to
avoid overwriting an existing file on the guest by appending an integer
to the filename. Instead of just appending the integer to the very end,
we now try to be smarter and append the integer before the file
extension. For example, if we transferred the file "file.doc" twice, the
second copy would have become "file.doc (1)". It is now "file (1).doc".
This matches more closely with the behavior of the Nautilus file manager
as well.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>