systemd: use "After" with "Requisite"
The "Requisite" directive is intended to require that the virtio port
device exists before starting the vdagent socket unit. If the device
does not exist, the socket should fail to start. However, without using
"After", the socket will still start even when the device is not
present.

The man page for systemd.unit (v239+) says the following about
'Requisite=':

"Similar to Requires=. However, if the units listed here are not
started already, they will not be started and the starting of this unit
will fail immediately. Requisite= does not imply an ordering
dependency, even if both units are started in the same transaction.
Hence this setting should usually be combined with After=, to ensure
this unit is not started before the other unit."

See also https://github.com/systemd/systemd/issues/4756 for further
details (particularly the following comment:
https://github.com/systemd/systemd/issues/4756#issuecomment-267005324)

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Acked-by: Lukáš Hrázký <lhrazky@redhat.com>