spice-vdagent.desktop: Change autostart to WindowManager stage
It is possible that g_get_user_special_dir() return NULL for
G_USER_DIRECTORY_DESKTOP or G_USER_DIRECTORY_DOWNLOAD; This is a sure
thing to happen in case xdg-user-dirs hasn't been run yet.

Keeping spice-vdagent to run on Initialization stage of
X-GNOME-Autostart-Phase could trigger a race with xdg-user-dirs that
also runs on Initialization, which is the earliest stage possible.

Moving to spice-vdagent to run WindowManager stage avoids the race
with xdg-user-dirs and still runs before any graphical application.

Tested also moving to Desktop stage, which is 4th stage (out of 5) and
worked without apparent drawbacks.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1623947
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Uri Lublin <uril@redhat.com>