Send device info to agent when it connects
Previously, the device info was only sent from the daemon to the session
agent when we receive the device info message from the host. That means
that if the session agent exits and reconnects, the session agent will
not have the device info necessary to translate spice display IDs to
guest device outputs. This patch saves the last device info message that
was received from the host, and re-sends it to the session agent when it
reconnects.

NOTE: this may not be necessary if the server simply re-sends a device
info message when it detects that the agent has connected...

Acked-by: Lukáš Hrázký <lhrazky@redhat.com>
Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>