gitlab-ci: run singular job for build options
There are two main bottlenecks in CI:
1. Network might be slow for bootstrapping;
2. Big job queues to get a runners to run our jobs.

This patch reduces the three jobs to a single one in order to prevent
long delays when CI resources are scarce which is a common scenario.

No major losses in log should happen. We might add specific .log
files as artifacts if we need them later on.

For completeness, use git clean -xfd for a clean git tree instead of
make distclean.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>