Add a non-X11 function to send the guest resolution to the daemon/server.
The purpose here is to further separate X11-specific code, and have the
actual sending to the daemon to be out of X11 code, and reusable by
other APIs.

- Replace vdagent_x11_send_daemon_guest_res() with a function that
builds
  and returns the list of monitor resolutions.
- Create the vdagent_display_send_daemon_guest_res() that uses this to
  send the resolutions to the daemon.
- Call this function in place of vdagent_x11_send_daemon_guest_res()

In later commits, we will be able to use a different API to generate
the list of monitor resolution, without having to change the X11 code.

Signed-off-by: Julien Ropé <jrope@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>