Split the "lookup_xrandr_output_for_device_info()" function to reuse some of it with different backends.
Most display drivers will number their outputs starting with "1", but
some (like older QXL drivers) will start from 0.
As we need to map device names with SPICE display IDs, we need to know
what to expect.

The logic that finds whether we have a O-based index or not is
independent from the backend, and can be reused outside of the X11
implementation.

This commit separates this logic for that purpose, to prevent code
duplication.
- create the has_zero_based_display_id() function to call whatever the
  backend, and learn whether our IDs should be decremented or not.
  Its code is taken out of lookup_xrandr_output_for_device_info()
- keep the lookup_xrandr_output_for_device_info() for x11 implementation
- add parameters to some functions to give them the information that IDs
  should be decremented because of the 0-based index.

Signed-off-by: Julien Ropé <jrope@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>