x11: move code to clipboard_data_send_to_requestor()
With the upcoming patch, clipboard data will be sent to the
requestor outside of vdagent_x11_clipboard_data() so
extract the necessary code into a new function to avoid
repetition.

The function takes "take_ownership" argument. If set to True,
the caller doesn't have to free the passed data which will
also be useful with the following changes.

Improve style by adding brackets to single-line if statements.

Signed-off-by: Jakub Janků <jjanku@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>