clipboard: enable copying files from client using webdav
When the user wants to copy files, new spice-gtk can share those
files using the existing phodav server. In that case, it advertises
the VD_AGENT_CLIPBOARD_FILE_LIST type in the clipboard grab message.

Upon request of the clipboard data in the mentioned type, spice-gtk
provides a list of absolute paths in the phodav server - these are
the files that are supposed to be copied/moved.

The role of the vdagent is to:
1) ensure that the phodav share is mounted,
2) adjust the provided paths given the mountpoint,
3) provide the uri list in various formats depending on which one
   was requested (different file managers use differenct formats)

The code that accomplishes these tasks is located in a new file
webdav-cb.c - the main reason for it is that vdagent currently
supports two clibpoard backends: x11, gtk+.

Implement this new feature only in the x11 backend since
the future of the gtk+ one is not clear.

Copy and move was tested with GNOME Nautilus, KDE Dolphin,
Cinnamon Nemo, Mate Caja, Xfce Thunar, LXDE PCManFM, Krusader.
The functionality with other file managers might be limited.

Copying files from the vdagnet side to the client is not
supported yet.

Signed-off-by: Jakub Janků <jjanku@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>